sources/frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Ken Harlbar <balotech514@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-fasthtml = "^0.4.4"
pymongo = {version = "3.12", extras = ["srv"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
